Removing a confusing (and I think unecessary) phrase.
I still dit not understand "in the same session will run ok, and not
erase it." In the issue you explained about xauth, but not about your
statement. It's strange to say "the container will run OK if you do not
erase it", since if you "erase" the container, of course it will no be
running OK (it will be erased). Sorry to be boring about that, but this
container you created is amazing, and if it's used by people do not
understand Docker (like me) the Readme can be really confusing. Anyway,
considering the text in #1, maybe my PR is OK.