Handle Locales according to Debian Docs
According to [1]:

"Using LC_ALL is strongly discouraged as it overrides everything. Please
use it only when testing and never set it in a startup file. "

And according to [2]:

"First, you have to set EnvironmentVariables such as LANG, LANGUAGE,
LC_CTYPE, LC_MESSAGES to your local language. Usually LANG (or LC_ALL)
is sufficient."

So I've left only LANG env variable.

Also, defining the variable in the docker-compose may not be enought, if
one builds the Dockerimage using docker (and not docker-compose) we
won't have the env variable available in order to build the locale while
building the image.

At last but not least, using localedef is a manual way to build the
locale. If at any point the locales are built again, then the manual
ones will be dropped, so it is prefered to use locale-gen.

1: https://wiki.debian.org/Locale
2: https://wiki.debian.org/ChangeLanguage